refactor(svelte2tsx): 単純ノードのハンドラを nodes/ へ切り出す
text / comment / mustache_tag / raw_mustache_tag / const_tag /
declaration_tag / debug_tag / render_tag / attach_tag を
上流 htmlxtojsx_v2/nodes/*.ts と 1 対 1 になるよう純粋移動。
出力・ロジックの変更なし。
