+++ crates/rsvelte_core/src/svelte2tsx/template/nodes/mod.rs
@@ -1,0 +1,11 @@
+//! Per-node-type template handlers, mirroring `htmlxtojsx_v2/nodes/*.ts`.
+
+pub(super) mod attach_tag;
+pub(super) mod comment;
+pub(super) mod const_tag;
+pub(super) mod debug_tag;
+pub(super) mod declaration_tag;
+pub(super) mod mustache_tag;
+pub(super) mod raw_mustache_tag;
+pub(super) mod render_tag;
+pub(super) mod text;
